requirements.txt
uiautomator2
pymysql
pytest
allure-pytest==2.6.2
allure-python-commons==2.6.2
redis